+++ src/tools.js
@@ -79,6 +79,16 @@ const bulletproofDecodeURIComponent = (encodedURIComponent) => {
     }
 };
 
+// Ports returned by `parseUrl` when port is not explicitly specified.
+// Values are based on node docs: https://nodejs.org/api/url.html#url_url_port
+const STANDARD_PORTS_BY_PROTOCOL = {
+    'ftp:': 21,
+    'http:': 80,
+    'https:': 443,
+    'ws:': 80,
+    'wss:': 443,
+};
+
 /**
  * Parses a URL using Node.js' `new URL(url)` and adds the following features:
  *  - `port` is casted to number / null from string
@@ -126,6 +136,11 @@ export const parseUrl = (url) => {
         if (matches && matches.length === 2) {
             parsed.scheme = matches[1];
         }
+    }
+
+    // Add default port based on protocol when no port is explicitly specified.
+    if (parsed.port === null) {
+        parsed.port = STANDARD_PORTS_BY_PROTOCOL[parsed.protocol] || null;
     }
 
     return parsed;